_04_excel_automation/package-requirements.txt
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
gspread==6.1.2
idna==3.7
oauthlib==3.2.2
openpyxl==3.1.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
urllib3==2.2.2
